Don't crash when we call __qualname__ on torch._C.ScriptFunction
We've root caused this to correctly throwing attribute error on ScriptFunction
when missing attributes are caused. This PR will fix crashes that are showing
up. I'm going to stack a second PR to fix torch._c.ScriptFunction just being a
very badly behaving python object (which should also fix this

ghstack-source-id: 51ddf6aaec309c1a777f8796298d93543c011f10
Pull Request resolved: https://github.com/pytorch/pytorch/pull/147894